# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: ROI Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /investments:
    get:
      operationId: getMyInvestments
      tags: [investments]
      summary: Get current user's investments
      responses:
        "200":
          description: List of investments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Investment"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createInvestment
      tags: [investments]
      summary: Create a new investment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvestmentRequest"
      responses:
        "201":
          description: Investment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/investments:
    get:
      operationId: adminGetInvestments
      tags: [admin]
      summary: Admin - Paginated investments with filters
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, completed, all]
        - name: userId
          in: query
          required: false
          description: Filter by Firebase user id
          schema:
            type: string
        - name: search
          in: query
          required: false
          description: Case-insensitive match on user email, name, or user id substring
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        "200":
          description: Paginated investments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInvestmentListPage"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: adminCreateInvestment
      tags: [admin]
      summary: Admin - Create a new investment for a user (multi-plan supported)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateInvestmentRequest"
      responses:
        "201":
          description: Investment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInvestment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User or plan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/investments/{investmentId}:
    patch:
      operationId: adminPatchInvestment
      tags: [admin]
      summary: Admin - Set manual investment status or override systemActive
      parameters:
        - name: investmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPatchInvestmentRequest"
      responses:
        "200":
          description: Updated investment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInvestment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Investment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        confirmPassword:
          type: string
          minLength: 6
          description: Must match password; validated server-side and not stored.
        phone:
          type: string
          description: Mobile number; digits only are stored after normalization (10–15 digits).
      required: [name, email, password, confirmPassword, phone]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [user, admin]
        walletBalance:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, phone, role, walletBalance, isActive, createdAt]
...
    Investment:
      type: object
      properties:
        id:
          type: string
        planId:
          type: string
        planName:
          type: string
        amount:
          type: number
        dailyRoi:
          type: number
        maxReturn:
          type: number
        totalEarned:
          type: number
        daysCompleted:
          type: integer
        maxDays:
          type: integer
        systemActive:
          type: boolean
        manualStatus:
          type: string
          enum: [active, inactive]
        isActive:
          type: boolean
        status:
          type: string
          enum: [active, completed, manually_stopped]
        startDate:
          type: string
          format: date-time
        lastRoiUpdate:
          type: string
          format: date-time
          nullable: true
      required: [id, planId, planName, amount, dailyRoi, maxReturn, totalEarned, daysCompleted, maxDays, systemActive, manualStatus, isActive, status, startDate]
    AdminInvestment:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        userEmail:
          type: string
        planId:
          type: string
        planName:
          type: string
        amount:
          type: number
        dailyRoi:
          type: number
        maxReturn:
          type: number
        totalEarned:
          type: number
        daysCompleted:
          type: integer
        maxDays:
          type: integer
        systemActive:
          type: boolean
        manualStatus:
          type: string
          enum: [active, inactive]
        isActive:
          type: boolean
        status:
          type: string
          enum: [active, completed, manually_stopped]
        startDate:
          type: string
          format: date-time
        lastRoiUpdate:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, userName, userEmail, planId, planName, amount, dailyRoi, maxReturn, totalEarned, daysCompleted, maxDays, systemActive, manualStatus, isActive, status, startDate]
    AdminInvestmentListPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminInvestment"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required: [items, total, page, pageSize, totalPages]
    AdminCreateInvestmentRequest:
      type: object
      properties:
        userId:
          type: string
        planId:
          type: string
      required: [userId, planId]
    AdminPatchInvestmentRequest:
      type: object
      properties:
        manualStatus:
          type: string
          enum: [active, inactive]
        systemActive:
          type: boolean
      description: At least one field should be provided
...
    CreateInvestmentRequest:
      type: object
      properties:
        planId:
          type: string
      required: [planId]
...