# lib/api-zod/src/generated/types/createInvestmentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ROI Investment Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface CreateInvestmentRequest {
  planId: string;
}